move slicedReducer into feature-u-redux pkg
