+++ src/feature/currentView/state.js
@@ -1,5 +1,5 @@
 import {reducerHash}      from 'astx-redux-util';
-import {slicedReducer}    from '../../util/feature-u';
+import {slicedReducer}    from '../../util/feature-u/aspect/feature-u-redux';
 import featureName        from './featureName';
 import actions            from './actions';
