+++ src/util/feature-u/aspect/feature-u-redux/reducerAspect.js
@@ -9,14 +9,21 @@ import isFunction         from 'lodash.isfunction';
 /**
  * ?? document ... somewhat dup of readme
  */
-export default createAspect({
-  name: 'reducer', // to fully manage all of redux, we ONLY need the reducers (hence our name)!
-  validateFeatureContent,
-  assembleFeatureContent,
-  assembleAspectResources,
-  getReduxStore,
-  injectRootAppElm,
-});
+
+// ?? problem in in-line expansion
+// src/util/feature-u/aspect/feature-u-redux/reducerAspect.js inline: createAspect(...)
+// src/util/feature-u/createAspect.js  invoking: isBuiltInFeatureKeyword()
+// src/util/feature-u/createFeature.js <<< prob: hasn't yet initialized KJB I think because the createAspect is kicking off too soon DUE TO pkging of feature-u-redux
+// ??? for now PUNT, and no-op the execution of createAspect() in reducerAspect
+export default 'NOT READY YET';
+// ? export default createAspect({
+// ?   name: 'reducer', // to fully manage all of redux, we ONLY need the reducers (hence our name)!
+// ?   validateFeatureContent,
+// ?   assembleFeatureContent,
+// ?   assembleAspectResources,
+// ?   getReduxStore,
+// ?   injectRootAppElm,
+// ? });
 
 
 /**
@@ -228,7 +235,7 @@ export function accumAppReducer(aspectName, activeFeatures) { // ... named expor
  *
  * @private
  */
-accumReducer(genesisNode) {
+function accumReducer(genesisNode) {
 
   if (isFunction(genesisNode)) {
     return genesisNode;
