+++ src/util/feature-u/index.js
@@ -1,5 +1,4 @@
 import createFeature     from './createFeature';
-import slicedReducer     from './slicedReducer';
 import managedExpansion  from './managedExpansion';
 import runApp            from './runApp';
 import createRoute       from './createRoute';
@@ -15,7 +14,6 @@ import createRoute       from './createRoute';
 //         import * as FeatureU from 'feature-u';
 export {
   createFeature,
-  slicedReducer,
   managedExpansion,
   createRoute,
   runApp,
@@ -28,7 +26,6 @@ export {
 //         const FeatureU   = require('feature-u');
 export default {
   createFeature,
-  slicedReducer,
   managedExpansion,
   createRoute,
   runApp,
